Now the AboutView is shown if login succeeds
To make this happen we had to refactor so that the AuthService
became a Singleton. It was achieved using the Locator (part of
ReactiveUI). Then the AuthService was injected into both
the MainWindowViewModel and LoginViewModel, so when the
login succeded via the LoginViewModel an event would be raised
by the AuthService that the MainWindowViewModel was listening to.
On that event the MainWindowViewModel swap to the AboutViewModel.