Showing live data
Implemented a LiveDataService that simulates live data. And is
started at the beginning of the application and stopped when
the application is shutting down.

We also created a LiveDataViewModel that has the service as a
dependency injection. We had to use the IActivatableViewModel
to contol the subscription and unsubscription of events to
avoid memory leaks.